+++ back-end/src/main/java/com/group_7/backend/entity/SmokingProfile.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "SmokingProfiles")
 @Getter
@@ -16,20 +18,20 @@ public class SmokingProfile {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "SmokingProfileID")
-    private long smokingProfileId;
+    private Long smokingProfileId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "UserID", referencedColumnName = "UserID", unique = true)
     private User user;
 
     @Column(name = "CigarettesPerDay")
-    private Integer cigarettesPerDay;
+    private int cigarettesPerDay;
 
     @Column(name = "CostPerPack")
-    private Double costPerPack;
+    private BigDecimal costPerPack;
 
     @Column(name = "WeekSmoked")
-    private Integer weekSmoked;
+    private int weekSmoked;
 
     @Column(name = "Note", columnDefinition = "NVARCHAR(MAX)")
     private String note;
